# backendApi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Punch System API",
        "description": "打卡系統 API 文件",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/positions": {
            "get": {
                "tags": [
                    "Position"
                ],
                "summary": "取得所有職位",
                "description": "返回所有職位列表，包含部門名稱",
                "operationId": "getAllPositions",
                "responses": {
                    "200": {
                        "description": "成功獲取職位列表",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "成功獲取所有職位"
                                        },
                                        "positions": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Position"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未授權",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "未授權的請求"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Position"
                ],
                "summary": "新增職位",
                "description": "建立一個新的職位",
                "operationId": "createPosition",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "專案經理"
                                    },
                                    "department_id": {
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "職位新增成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "職位新增成功"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "驗證失敗",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "職位名稱已存在"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/positions/by/department/{name}": {
            "get": {
                "tags": [
                    "Position"
                ],
                "summary": "根據部門篩選職位",
                "description": "根據部門名稱篩選職位",
                "operationId": "getPositionsByDepartment",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "部門名稱",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "資訊部"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功獲取職位",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "department": {
                                            "type": "string",
                                            "example": "資訊部"
                                        },
                                        "positions": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Position"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "找不到該部門",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "找不到該部門"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Position"
                ],
                "summary": "為部門指派職位",
                "description": "將職位指派到特定部門",
                "operationId": "assignPositionToDepartment",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "部門名稱",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "資訊部"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "職位已指派到部門",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "職位已指派到部門"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "找不到該部門",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "找不到該部門"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/positions/{id}": {
            "delete": {
                "tags": [
                    "Position"
                ],
                "summary": "刪除職位",
                "description": "刪除指定職位",
                "operationId": "deletePosition",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "職位 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "職位刪除成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "職位刪除成功"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "找不到職位",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "找不到職位"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Position"
                ],
                "summary": "更新職位",
                "description": "更新職位名稱或所屬部門",
                "operationId": "updatePosition",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "職位 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "資深開發工程師"
                                    },
                                    "department_id": {
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "職位更新成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "職位更新成功"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "找不到職位",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "找不到職位"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Department": {
                "title": "Department",
                "description": "部門資料結構",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "人事部"
                    }
                },
                "type": "object"
            },
            "Position": {
                "title": "Position",
                "description": "職位資料結構",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "資深工程師"
                    },
                    "department_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "department": {
                        "$ref": "#/components/schemas/Department"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}